Improve X API tweet fidelity: note_tweet text and t.co URL expansion
Roast quality dropped after switching to the X API because a third of the
scraped tweets were opaque https://t.co/... tokens: long posts truncated
(no note_tweet) and links unexpanded. Now the full long-form text is used
and t.co links are replaced with their expanded URLs, so the model sees
real material again.
